requirements-dev.txt
aiokafka==0.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
beartype==0.22.9
boto3==1.42.30
boto3-stubs==1.42.30
botocore==1.42.30
botocore-stubs==1.42.30
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
defusedxml==0.7.1
execnet==2.1.2
fastapi==0.127.1
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.7.8
lxml==6.0.2
mando==0.7.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
moto==5.1.20
multidict==6.7.0
mypy==1.19.1
mypy-boto3-s3==1.42.21
mypy-extensions==1.1.0
packaging==25.0
pathspec==1.0.3
pip==25.3
pluggy==1.6.0
py-partiql-parser==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyld==2.0.4
pymysql==1.1.2
pyparsing==3.3.1
pysocks==1.7.1
pytest==8.4.2
pytest-asyncio==0.26.0
pytest-cov==4.1.0
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
radon==6.0.1
rapidhash==0.1.0
rdflib==7.5.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
responses==0.25.8
rich==14.2.0
rpds-py==0.30.0
ruff==0.14.13
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
types-awscrt==0.31.1
types-boto3==1.42.30
types-jsonschema==4.26.0.20260109
types-pymysql==1.1.0.20251220
types-pyyaml==6.0.12.20250915
types-requests==2.32.4.20260107
types-s3transfer==0.16.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
vulture==2.14
werkzeug==3.1.5
xmltodict==1.0.2
